+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Added missing building argument in TAC state
 - PCT calculation logic update.
 
+## [1.2.0] - 2021-12-31
+
+### Features
+- Introduced "Cost Threshold" in the AdminToolkit, where we can set the margin of the cost,
+  So that in the TAC only certain users can move the project from one state to another.
+
 ## [1.1.0] - 2021-12-31
 
 ### Features